Improve Compatibility with Firefox/Gecko
(1) Firefox/Gecko requires that the popup's width be defined in the
body, not the root, so width was defined in the body.

(2) Firefox/Gecko scrollbars are less customizable, but I was able to
approximate the scrollbar appearance in Chrome/Edge by adding
"scrollbar-color" and "scrollbar-width" properties.

(3) Moved the ::-webkit selectors to the bottom.