Improve Compatibility with Firefox/Gecko
(1) Recreated the hide/unhide animation. The "Hidden" indicator is now a
sibling to the toggle-button-text, rather than an ::after
pseudo-element.
The pseudo-element animation depended on the toggle-button-text element
having an overflow value of "clip visible", which rendered as intended
in Chrome/Edge, but not in Firefox.

(2) Added a title attribute to the toggleButtonTextElement for cases
when a long company name has hidden overflow.