Announce Content Script Status
(1) The content script now announces informs the background script and
popup script when there are changes to the Hide n' Seek UI, i.e. whether
or not there are .job-block-elements in the DOM. This approach works
better on single page application (SPA) websites, since they do not
reload the DOM while navigating.

(2) The JobRegistrar mutation observer has been simplified. Instead of
searching through all record targets and added nodes for new jobs,
it simply performs a querySelectorAll during each callback. So far, I
have not noticed any performance issues, but if there are in the future,
I can debounce the callback.

(3) The mutation observer callback also sends a message to the
background script and popup script if there are any changes to the
Hide n' Seek UI.