Add Support for Signed-Out LinkedIn Job Search
(1) Until recently, I was unaware that LinkedIn had a searchable job
board for users without a LinkedIn account. The signed-out version of
LinkedIn's job search page has markup and classes that differ from the
signed-in version, and these differences are now accounted for, allowing
Hide n' Seek users to hide jobs whether or not they are seeing the
signed-in or signed-out version of LinkedIn's job search page.

(2) A new class, JobBoardSelectors was created to better organize all
selectors needed by other classes.

(3) JobBlockElementInserter class constructor now takes an ancestor
selector as a parameter and uses it to create the insertion reference
element. It also sets "position: relative" on the parent element of the
insertion reference element. This is done because the job block element
has an absolute position.

(4) JobElementDataGetter class was renamed JobAttributeValueGetter since
it now returns a single value and not a data object. Logic was also
rewritten to be more generalizable and less repetitive.