Remove chrome.tabs.onActivated Listener
(1) Removed the chrome.tabs.onActivated listener. Previously, it was
used to update the extension badge and text, but this function is not
necessary. The listeners and callbacks for chrome.tabs.onUpdated and
chrome.storage.local.onChanged are sufficient.

(2) JobBoards.getJobBoardIdByUrl validates url just in case.

(3) The previously-added sendResponses have been removed. They were
unnecessary and were only added in an attempt to solve the nonresponsive
listeners bug. https://stackoverflow.com/a/75912240/21288253