Optimize Badge Updating Logic
(1) Previously, all badges for all tabs with a given jobBoardId would be
updated whenever a content script notified the background script of a
change in the Hide n' Seek UI. However, updating the badges of other
tabs was unnecessary work. Now, when a content script sends this
message, the background script will only update the badge for that
specific tab.

(2) JobBoards.getContentScriptStatusesOfJobBoardIdTabs method was
removed since updateBadgeTextAndTitle now takes tabs as an argument
instead of automatically getting and updating all jobBoardId tabs.