Add Connectivity Handling
(1) The job search popup now listens for changes in connectivity and
will automatically enable or disable inputs depending on whether the
user's device is online or offline. If the user's device is offline,
Hide n' Seek will inform the user by displaying the message "Device is
offline" inside the search input field.

(2) When a user searches for a job with the popup, Hide n' Seek will now
first check that the job board site is online or not. If the job board
site is offline, Hide n' Seek will inform the user by displaying a
message (e.g. "Can't connect to Indeed") inside a flashing-red search
input field.