Optimize for bfcache
(1) Using the back/forward buttons on Indeed triggered the use of
bfcache, which would prevent the extension button's badge test from
updating. Instead of showing the number of hidden jobs, there was
nothing. This issue was resolved by adding a pageshow listener and
checking the pageTransitionEvent.persisted property. Now, when bfcache
is used (persisted === true), the content script notifies the background
and popup scripts so that the badge can be updated.