Overhaul storage syncing strategy
(1) Added storage chunking and dechunking mechanisms. In order to make better use of sync storage space and not exceed the per-item quota, storage is now chunked before setting on sync storage. This now allows users to block thousands of jobs instead of just hundreds.

(2) Changed the onInstall listener so that it only copies sync storage so local storage on install, not on updates, etc.

(3) Replaced syncId generator with crypto.randomUUID

(4) Overhauled and simplified the syncing mechanism, which no longer stores or depends on a list of pending changes.