Revise LinkedIn company name getter
LinkedIn's job listing page structure and element features vary depending on whether you're signed in or not, chaining querySelector("span") would lead to issues matching listings on the non-signed-in version of LinkedIn's job listings, which doesn't put the company name in a element. So, instead of chaining a querySelector, I changed part of the selector list from ".artdeco-entity-lockup__subtitle" to ".artdeco-entity-lockup__subtitle > span" so that it only applies to that particular selector used in the signed-in version of LinkedIn. If you have time, can you check that this change resolves the conflict with inDoors? I installed inDoors but it doesn't seem to be working for me, so I wasn't able to reproduce the issue.